requirements_local.txt
imutils==0.5.3
Keras==2.4.3
tensorflow==2.4.0
numpy==1.19.3
opencv_python==4.4.0.46
Pillow==8.2.0
scikit_learn==0.24.2
h5py~=2.10.0
flask==1.1.2